+++ tests/testthat/test-as.list.R
@@ -20,3 +20,13 @@ test_that("cdata nodes become character vectors", {
 test_that("xml attributes become R attibutes", {
   expect_equal(list_xml("<x a='1' b='2'></x>"), structure(list(), a = "1", b = "2"))
 })
+
+test_that("xml names are preserved when attributes exist", {
+  expect_equal(list_xml("<x a='1' b='2'><y>3</y><z>4</z></x>"),
+               structure(list(y = list('3'), z = list('4')), names = c("y", "z"), a = "1", b = "2"))
+})
+
+test_that("special attributes are escaped", {
+  expect_equal(list_xml("<x a='1' b='2' names='esc'><y>3</y><z>4</z></x>"),
+               structure(list(y = list('3'), z = list('4')), names = c("y", "z"), a = "1", b = "2", .names='esc'))
+})